requirements.txt
flask
py-cpuinfo
psutil
python-nmap